Setup instructor routes
