+++ frontend/src/views/Layout/Layout.tsx
@@ -36,6 +36,7 @@ const menuItemsForRole: Record<UserRole, MenuItem[]> = {
     { text: 'Ogłoszenia', icon: 'campaign', link: '/ogloszenia' },
   ],
   [UserRole.Instructor]: [
+    { text: 'Moje Jazdy', icon: 'toys', link: '/moje-jazdy' },
     { text: 'Kursanci', icon: 'school', link: '/kursanci' },
     { text: 'Instruktorzy', icon: 'class', link: '/instruktorzy' },
     { text: 'Pojazdy', icon: 'directions_car', link: '/pojazdy' },
