+++ frontend/src/views/Lessons/TraineeMyLessons/LessonsCalendar/LessonsCalendar.utils.ts
