+++ frontend/src/views/Lessons/TraineeMyLessons/TraineeMyLessons.styled.ts
